education/requirements.txt
PyYAML~=6.0
pandas~=2.0.2
fastapi~=0.110.1
pydantic~=2.3.0
starlette~=0.37.2
h11~=0.14.0
pip~=23.2.1
attrs~=22.2.0
wheel~=0.40.0
Pillow~=9.5.0
tornado~=6.3.2
Jinja2~=3.1.2
filelock~=3.13.3
pytz~=2022.7.1
sniffio~=1.3.0

numpy~=1.23.5
setuptools~=65.5.0
overrides~=7.3.1
cffi~=1.15.1
future~=0.18.3
click~=8.1.7
ipython~=8.14.0
SQLAlchemy~=2.0.22
matplotlib~=3.7.1
MarkupSafe~=2.1.3
scipy~=1.10.1
python-dateutil~=2.8.2
requests~=2.28.2
websockets~=12.0
anyio~=3.7.0
uvicorn~=0.29.0
httptools~=0.6.1
watchfiles~=0.21.0
six~=1.16.0
greenlet~=3.0.0
psutil~=5.9.5
decorator~=4.4.2
python-multipart~=0.0.9
packaging~=23.1
zipp~=3.18.1
build~=1.2.1
platformdirs~=4.2.0